UmlComponent, corrections & javadoc
