+++ src/model/UmlEntity.java
@@ -3,11 +3,18 @@ package model;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.Set;
-
+/**
+ * Abstract class, an UML entity, parent of UmlComponent
+ * @see UmlComponent
+ * @author christian, fabien, bastien
+ *
+ */
 public abstract class UmlEntity{
-
+	
+	// Attributes
+	
 	/**
-	 * Visibilité d'un élément
+	 * Visibility of an element
 	 */
 	private Visibility visibility;
 
@@ -16,51 +23,88 @@ public abstract class UmlEntity{
 	 */
 	private HashSet<Modifier> modifiers;
 	
+	// Constructor
+	
 	/**
-	 * Construire un élément avec une visibilité et
-	 * des modifieurs
+	 * Constructor of an element with visibility and modifier(s)
 	 * @param visibility the visibility
-	 * @param modifier the method modifieur(s)
+	 * @param modifier the method modifier(s)
 	 */
 	public UmlEntity(Visibility visibility, Collection<Modifier> modifier) {
 		this.visibility = visibility;
 		this.modifiers = new HashSet<Modifier>(modifier);
 	}
 	
+	/**
+	 * Constructor of an element with visibility
+	 * @param visibility the visibility
+	 */
 	public UmlEntity(Visibility visibility) {
 		this.visibility = visibility;
 		this.modifiers = new HashSet<Modifier>();
 	}
 	
+	/**
+	 * Constructor of an element by default
+	 */
 	public UmlEntity() {
 		this.visibility = Visibility.PUBLIC;
 		this.modifiers = new HashSet<Modifier>();
 	}
 	
+	// Methods
+	
+	/**
+	 * Add a modifier to the modifiers set
+	 * @param modifier a modifier of the entity
+	 */
 	public void addModifier(Modifier modifier) {
 		this.modifiers.add(modifier);
 	}
 	
+	/**
+	 * Clear the modifiers set
+	 */
 	public void clearModifiers() {
 		this.modifiers.clear();
 	}
 	
+	/**
+	 * Remove a modifier from the modifiers list
+	 * @param modifier the modifier to remove
+	 */
 	public void removeModifier(Modifier modifier) {
 		this.modifiers.remove(modifier);
 	}
 	
+	/**
+	 * Getter visibility
+	 * @return visibility the visibility
+	 */
 	public Visibility getVisibility() {
 		return this.visibility;
 	}
 	
+	/**
+	 * Setter visibility
+	 * @param visibility the visibility
+	 */
 	public void setVisibility(Visibility visibility) {
 		this.visibility = visibility;
 	}
 	
+	/**
+	 * Getter set of modifiers
+	 * @return Set<Modifier> the set of modifiers
+	 */
 	public Set<Modifier> getModifier() {
 		return this.modifiers;
 	}
 	
+	/**
+	 * Setter set of modifiers
+	 * @param modifiers a collection of modifiers
+	 */
 	public void setModifiers(Collection<Modifier> modifiers) {
 		this.modifiers = new HashSet<Modifier>(modifiers);
 	}
